tests: Move all fake devices to common directory
