+++ tests/fake_device/fake_mrp_atv.py
@@ -10,7 +10,7 @@ from pyatv.mrp import chacha20, messages, protobuf, variant
 from pyatv.mrp.protobuf import CommandInfo_pb2 as cmd
 from pyatv.mrp.protobuf import SetStateMessage as ssm
 from pyatv.mrp.server_auth import MrpServerAuth
-from tests.airplay.fake_airplay_device import FakeAirPlayDevice, AirPlayUseCases
+from tests.fake_device.fake_airplay_device import FakeAirPlayDevice, AirPlayUseCases
 
 _LOGGER = logging.getLogger(__name__)
