+++ tests/mrp/test_mrp_auth.py
@@ -7,7 +7,7 @@ from pyatv import exceptions
 from pyatv.const import Protocol
 from pyatv.conf import MrpService, AppleTV
 from pyatv.mrp.server_auth import PIN_CODE, CLIENT_IDENTIFIER, CLIENT_CREDENTIALS
-from tests.mrp.fake_mrp_atv import FakeAppleTV
+from tests.fake_device.fake_mrp_atv import FakeAppleTV
 
 
 class MrpAuthFunctionalTest(AioHTTPTestCase):
